Create `...TestEnv` classes
